It's not our place to police inputs
Removing policing of input values because the user could change whether or not any given field is required at any time they choose and this code wouldn't obey it. It means we can't be 100% certain that every register event is a successful register, but it also means we're not going to block registrations accidentally - that would be far worse.